+++ lerna.json
@@ -1,9 +1,6 @@
 {
   "version": "4.5.5",
-  "packages": [
-    "packages/*",
-    "examples/*"
-  ],
+  "packages": ["packages/*", "examples/*", "test-apps/*"],
   "npmClient": "yarn",
   "useWorkspaces": true,
   "useNx": true
